feature(order-product): Adicionar autorizacao em /app/order

+++ src/modules/app/services/OrderService.ts
@@ -1,4 +1,5 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { ICurrentUser } from 'modules/common/interfaces/currentUser';
 import { IOrder } from 'modules/database/interfaces/IOrder';
 import { Order } from 'modules/database/models/Order';
 import { OrderRepository } from '../repositories/OrderRepository';
@@ -13,18 +14,24 @@ export class OrderService {
     private productRepository: ProductRepository
   ) {}
 
-  public async remove(id: number) {
-    const order = await this.orderRepository.findById(id);
+  public async list(currentUser: ICurrentUser) {
+    return this.orderRepository.list(currentUser);
+  }
+
+  public async remove(id: number, currentUser: ICurrentUser) {
+    const order = await this.orderRepository.findById(id, currentUser);
     if (!order) throw new NotFoundException('not-found');
 
-    // To-Do: Checar se a ordem eh do usuario atual antes de remover
+    if (order.userId !== currentUser.id) throw new UnauthorizedException('Sem autorizacao para o recurso');
 
     return this.orderRepository.remove(id);
   }
 
-  public async save(model: IOrder): Promise<Order> {
+  public async save(model: IOrder, currentUser: ICurrentUser): Promise<Order> {
     const user = await this.userRepository.findById(model.userId);
     if (!user) throw new NotFoundException(`Usuario ${model.userId} nao encontrado`);
+
+    if (user.id !== currentUser.id) throw new UnauthorizedException('Sem autorizacao para o recurso');
 
     const order = await this.orderRepository.insert(model);
 
@@ -41,10 +48,11 @@ export class OrderService {
       throw e;
     }
 
-    return this.orderRepository.findById(order.id);
+    return this.orderRepository.findById(order.id, currentUser);
   }
 
-  public async update(model: IOrder) {
+  public async update(model: IOrder, currentUser: ICurrentUser) {
+    if (model.userId !== currentUser.id) throw new UnauthorizedException();
     return this.orderRepository.update(model);
   }
 }
